// mobile/src/api/generated/model/taskState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.40.2
 */

export type TaskState = typeof TaskState[keyof typeof TaskState];


export const TaskState = {
  open: 'open',
  closed: 'closed',
} as const;